+++ core/datatrak.c
@@ -17,6 +17,7 @@
 #include "file.h"
 #include "divesite.h"
 #include "ssrf.h"
+#include "tag.h"
 
 static unsigned int two_bytes_to_int(unsigned char x, unsigned char y)
 {
