+++ core/load-git.c
@@ -22,6 +22,7 @@
 #include "membuffer.h"
 #include "git-access.h"
 #include "qthelper.h"
+#include "tag.h"
 
 const char *saved_git_id = NULL;
