+++ core/save-html.c
@@ -8,6 +8,7 @@
 #include "qthelper.h"
 #include "gettext.h"
 #include "divesite.h"
+#include "tag.h"
 #include <stdio.h>
 
 void write_attribute(struct membuffer *b, const char *att_name, const char *value, const char *separator)
